src/delivery/requirements.txt
pulsar-client==2.10.1
pulsar-client[avro]==2.10.1
psycopg2-binary==2.9.5